+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 Version 1.7-SNAPSHOT
 -------------
 
+ADDED:
 
+- new class GenericClassFieldsExclusionStrategy with a class and one set of field names that have to be excluded
+- new class GenericMapClassFieldsExclusionStrategy with a map that defines the exclusion
 
 Version 1.6
 -------------